# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from "./uploadUrlRequest";

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Relative path for serving via GET /storage/public-objects/{objectPath} (e.g. catalog/uuid-filename.jpg). */
  objectPath: string;
  metadata?: UploadUrlRequest;
}